[sasl] Sort applications used by other applications correctly in .script
OTP-9984

Applications that are listed in {applications,Apps} in the app file
were not sorted correctly in the script file. They were loaded/started
in the reverse order of how they were listed in the .app file.

This is corrected so they are now sorted (internally between each
other) in the same way as they are listed in the .rel file.