Avoid undefined start_phases entry in .script
Inorder to align with reltool a {start_phases,undefined} entry in will
no longer occur in the application specification in .script files. The
start_phases entry will only exist if it contains real start phases.

Also, the default value for #application.start_phases is no longer set
to [] in systools.hrl. This value was actually never used, since it
was always explicitly set in systools_make:parse_applications - by
default to 'undefined', so the default value in the record definition
was only a confusing detail. Anyway - now corrected.