Eliminate prefetch for conditional instructions
Not pre-fetching in conditional instructions (instructions that use
-fail_action) seems to improve performance slightly.

The reason for that is that conditional instructions may jump to the
failure label, wasting the pre-fetched instruction. Another reason
is that that many conditional instructions do a function call, and
the register pressure is therefore high. Avoiding the pre-fetch
may reduce the register pressure and pontentially result in more
efficient code.