DTrace bug in message-send
dtrace doesn't print sender pid and receiver pid in message related
probe correctly, it truncates pid string to 7 characters on 64bit
platform(or 3 characters on 32bit platform).  The dtrace related code
did not set the string length correctly, it sets the string length to
the length of a pointer rather than the buffer length.