Remove stale code for hybrid heap and incremental GC
The hybrid heap emulator was last working in the non-SMP R11B
run-time system. When the constant pools were introduced in R12B,
the hybrid heap emulator was not updated to handle them.

At this point, the harm from reduced readability of the code is
greater than any potential usefulness of keeping the code.