erts: Make 64-bit installer look for MS Redistributable in the right places
The installer runs in 32-bit mode so 64-bit windows does not
show it the system directories a 64-bit application will see.
To find out if the MS C++ 2010 Redistributable Package is
installed the installer for 64-bit Erlang must search
for msvcr100.dll in:
  $WINDIR\sysnative\ $WINDIR\WinSxS\amd64*\
instead of for 32-bit Erlang
  $SYSDIR\ $WINDIR\WinSxS\x86*\

This patch also makes the installer if called in silent mode
call the MS Redistributable installer in silent mode.