ssl: Calculate handshake hash only when needed
TLS/SSL version before 1.2 always used a MD5/SHA combination
for the handshake hashes. With TLS 1.2 the default hash is
SHA256 and it is possible to negotiate a different hash.

This change delays the calculation of the handshake
hashes until they are really needed. At that point the hash
to use should be known.

For now MD5/SHA is still hard coded.