compile_SUITE: Correct the forms_2 test case to work on Windows
Correct the forms_2 test case introduced in 2d785c07fbf9f533bf so
that it will work on Windows. As originally written, the test case
assumed that filename:absname("/foo/bar") would return "/foo/bar",
which is not true on Windows (typically, the result will be
"c:/foo/bar").

While at it, clean up indentation, the overlong line, and comments.