Fix flattening of paths in erl_prim_loader
When correcting OTP-10071, a new error was introduced in
erl_prim_loader. In order to improve ability to detect if a file was
inside the primary archive, all paths were flattened - i.e. "."  and
".." were removed. This implementation had some faults, and it did not
take symlinks into account. This has been corrected.