beam_bool: Recognize more safe optimizations
The optimizer of boolean expressions can often reject optimizations
that it does not recognize as safe. For instance, if a boolean
expression was preceded by 'move' instructions that saved x registers
into y registers, it would almost certainly reject the optimization
because it required the y register not be used in the code that
follows.

Fix this problem by allowing identical 'move' instructions that assing
to y registers at the beginning of the old and the optimized code.

While at it, correct the spelling of "preceding".