v3_kernel: Fix match code for matched out segment size in multiple clauses
When matched variable is used as a size field in multiple clauses,
as in:

foo(<<L:8,A:L>>) -> A;
foo(<<L:8,A:L,B:8>>) -> {A,B}.

the match tree would branch out before the segment that used the
matched-out variable (in this example, the tree would branch out before
the matching of A:L). That happens because the pattern matching
compilator did not take variable substitutions into account when
grouping clauses that match the same value.

That is, the generated code would work similarly to this code:

foo(<<L:8,T/binary>>) ->
  case T of
    <<A:L>> ->
      A;
    _ ->
      case T of
        <<A:L,B:8>> ->   %% A matched out again!
	  {A,B}
	end
  end.

We would like the matching to work more like:

foo(<<L,A:L,T/binary>>) ->
  case T of
    <<>> -> A;
    <<B:8>> -> {A,B}
  end.

Fix the problem by taking the substitutions into account when grouping
clauses that match out the same value.