compiler: Warn if the size of a binary segment is invalid
The compiler would silently accept and Dialyzer would crash on
code like:

  <<X:(2.5)>>

It is never acceptable for Dialyzer to crash. The compiler should
at least generate a warning for such code. It is tempting to let
the compiler generate an error, but that would mean that code like:

  Sz = 42.0,
  <<X:Sz>>.

would be possible to compile with optimizations disabled, but not
with optimizations enabled.

Dialyzer crashes because it calls cerl:bitstr_bitsize/1, which
crashes if the type of size for the segment is invalid. The easiest
way to avoid that crash is to extend the sanity checks in v3_core
to also include the size field of binary segments. That will cause
the compiler to issue a warning and to replace the bad binary
construction with a call to erlang:error/1. (It also means that
Dialyzer will not issue a warning for bad size fields.)