make_preload: Don't fail if Perl's default file encoding is UTF-8
Setting Perl's default encoding for files to UTF-8, for example
like this:

  PERL_UNICODE=DS make

would crash the build with a message similar to:

  form size 1413 greater than size 1237 of module at
  utils/make_preload line 175, <FILE> chunk 1.

Tell Perl to interpret the data in BEAM files as binary by
using the binmode() function. The binmode() function existed
before Unicode support was added to Perl, which means that
make_preload should work even in old versions of Perl.

Noticed-by: Aaron Harnly