beam_jump: Don't move a block which can be entered via a fallthrough
beam_jump moves short code sequences ending in an instruction that causes
an exception to the end of the function, in the hope that a jump around
the moved blocked can be replaced with a fallthrough. Therefore, moving
a block that is entered via a fallthrough defeats the purpose of the
optimization.

Also add two more test cases for the beam_receive module to ensure that
all lines are still covered.