global: Allow "tuple funs" as resolve functions
Commit de7e01c958ff7c9e6da4034a53567a30a4ae5792 removed support for
tuple funs. To still allow resolve functions for
global:{re_}register_name/3 to be specified as {M,F}, we'll need to
translate {M,F} to an external fun.

Also update the documentation to mention that the use of {M,F} to
specify a resolve function is deprecated, and to add a recommendation
to use an external fun.