test_server_ctrl: Don't die in stop_extra_tools/1
If it is not possible for test_server_ctrl:start_log_file/0 to
create the log file directory or log files, then stop_extra_tools/1
(called from init_tester/10) is also likely to crash and hide the
original error reason. That will be confusing to the user.

Therefore, catch the call to stop_extra_tools/1.