[reltool] Keep order from rel spec when sorting used and included applications
The rel specification now dictates the order in which included and
used applications are loaded/started by the boot file. If the
applications are not specified in the rel spec, then the order from
the .app file is used. This was a bug earlier reported on systools,
and is now also implemented in reltool.

Example:

If a .app file specified

{applications,[x,y]}
{included_applications,[b,c]}

And the reltool.config has

{rel, "myrel", "1.0", [a,y,x,c,b]}

Then the boot file will load/start y before x and c before b. Earlier
x would always be started before y and b always before c due to the
order in the .app file.