common_test: Don't use undocumented features of io:format()
The Format string argument for io:format() is not documented
to accept an iolist, so we should not depend on it.

Also, it is bad practice to use an arbitrary string as a format
string for io:format(), since it could contain a '~' character
which could trigger a badarg exception.

Fix both problems at the same time by using io:put_chars() to
display the iolist.