Fix missing information in crash dump for native-compiled modules
In the erl_crash.dump file, native-compiled modules did not have
any information about attributes and compilation.

The problem is that the code:make_stub_module/3 BIF (which is
internally used when native code is loaded) did not copy the size
field the attribute and compilation info chunks. Those size fields
are only used when writing crash dumps.