test_server_SUITE: Remove the undefined_functions/1 test case
The undefined_functions/1 test case will fail if it is run
using an uninstalled Erlang/OTP system (i.e. in a source
tree/git repository).

Since the test_server application is now a part of Erlang/OTP
(it was not originally), the system-wide test case that looks
for use of undefined functions will catch undefined functions.

Therefore the easiest solution is to remove the test case.