[common_test] Extend timeout values for slave nodes in ct_master_SUITE
Some tests in ct_master_SUITE failed on some platforms with
boot_timeout when trying to start slave nodes. The default timeout
value was 3 seconds, and it is now extended to 10 seconds in these
tests.

To avoid similar errors with init_timeout and startup_timeout, these
are both exteded from 1 second default value to 10 seconds in this
test.