[test_server] Don't stop cover after test is finished
This commit removes all calls to cover:stop from test_server. This is
to avoid crash in some processes due to old code.

The correction shall be further refined for general use. There will,
for instance, be an option to turn off cover:stop/0 after completed
test run, so the default behaviour is backwards compatible.