beam_makeops: Eliminate a deprecation warning
Perl 5.16.1 (and perhaps other versions) issues the following
warning:

  defined(@array) is deprecated at utils/beam_makeops line 1714.
        (Maybe you should just omit the defined()?)

for the following line:

$prev_last = pop(@{$gen_transform{$key}})
	if defined @{$gen_transform{$key}};   # LINE 1714

The documentation for "defined" says that its use on hashes and
arrays is deprecated and that it may stop working in a future
release.

Simply removing "defined" (as suggested by the warning message)
will not work, as there will be an error when trying to use an
undefined value as an array reference:

  Can't use an undefined value as an ARRAY reference at
  utils/beam_makeops line 1714.

What we must do is to check whether $gen_transform{$key} is
defined before trying to use it as an array reference.

Noticed-by: Tuncer Ayaz