Keep the information about the current test case consistent
Three pieces of information could be out of sync in testcase
supervisor process at the time when a timetrap occurred:

* test_server_loc (process dictionary) - may indicate that a
  framework function is executing

* test_server_init_or_end_conf (process dictionary) - indicates whether
  init_per_testcase or end_per_testcase is executing

* The current configuration (#st.config) - set using a synchronous
  call

There could be in a crash in spawn_fw_call/7 because the current
configuration was not defined when it was expected to.

To avoid the problem, introduce set_tc_state/2 (and a corresponding
message) to allow setting both an indication what the testcase is
executing (e.g. init_per_testcase, framework call, and so on), and
the current configuration. Use only that information to handle
a timetrap timeout (and aborted testcase and the other reasons
for the testcase process to terminate). Completely remove
test_server_init_or_end_conf.