Improve surefire xml <testcase> element
Remove the 'description' attribute. It is not generated by Apache
Ant or Maven Surefire and probably ignored by all other Surefire
tools. Confirmed for Jenkins.

Add testcase description (if there is one) to the 'name' attribute.

Change format of 'name' attribute to look more like the output from
running eunit with verbose. Line number and arity was quite obscure
and arity is not useful as it is always 0.