[test_server] Add option {start_cover,false} to test_server:start_node
By default the test server will start cover on all nodes when the test
is run with code coverage analysis. To make sure cover is not started
on a new node, this option can be set.

This can be useful if the test itself starts cover or if the new node
for some reason can not run cover compiled code.