[test_server] Add Config to run_test_case_msgloop/1 from the start
Earlier, #st.config was set to undefined when entering this loop. Then
it was updated when the test case itself started. This means that if
the test case process crashed during test case initiation, the parent
process would not know the Config.

This commit saves the initial Config value in the parent process
(run_test_case_msgloop/1) from the start.