global_SUITE: Eliminate redundant registration of 'test_server' process
global_SUITE registers the name 'test_server' in the global registry
for what seems to be only historical reasons. Exactly how it happened
is lost in the mists of time, but I think the following is a reasonable
reconstruction:

* At one point in time, the test_server application started to do
  a global registration of a 'test_server' process to support
  embedded systems when the test_server application was run on both a
  host and a target system.

* global_SUITE had to be modified to take the pre-existing
  'test_server' registration into account.

* When Erlang/OTP switched to using common_test, global_SUITE was
  modified to register the 'test_server' name itself in commit
  8bd07f450816eaaae52d28740d89ead1f930cd02 (since the test_server
  application does not register the 'test_server' name when run
  under control of common_test).

To slightly reduce the size of global_SUITE, remove the redundant
registration and stop expecting 'test_server' to be registered.