Remove obsolete back-ends and simplify the options
It is time to clean up the mess of back-ends.

Remove all the obsolete back-ends and simplify the options used
to select them.

New Option  Old Equivalent
----------  --------------
ber         ber_bin,optimize,nif
per         per,optimize,nif
uper        uper_bin

The old options will still be recognized and translated to the
new options, but will also print a warning.

That implies that deprecated features that only are implemented
in the old 'ber' back-end will no longer work (e.g. the
{Typename,Value} notation).

Also make the return type for the generated encode/2 function
consistent. It used to be a binary for per and uper, and an iolist
for ber. Always make it a binary.