[common_test] Allow server to terminate netconf session before rpc-reply
OTP-10570

ct_netconfc:close_session sometimes returned {error,closed} because
the ssh connection was closed (from the server side) before the
rpc-reply was received by the client. This is normal and can not be
helped. It has been corrected so the return will be 'ok' in this case.

Other error situations will still give {error,Reason}.