Suppress warnings for unused variables using a compiler option
Currently, the generated code suppresses warnings for unusued variables
by assigning to the "_Val" variable, for example:

  _Val = Tmpval5

To be completely safe, that should have been:

  _ = Tmpval5

in case there happens to more than one such assignment in the same
scope.

However, a better way to suppress warnings is to simply use the
'nowarn_unused_vars' compiler option. That method will also work
for more complicated expressions there it would be cumbersome to
suppress warnings by assignment to "_".